# app/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API системы заявок",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Получить токен авторизации",
                "description": "Аутентификация пользователя и возврат токена доступа",
                "operationId": "1421b80e4f105abe77413791810d4270",
                "requestBody": {
                    "description": "Учетные данные",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная аутентификация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "a7d82K9vRqPwLb3cX1mZ0yT5nFgH6jSe"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные"
                    }
                }
            }
        },
        "/api/requests": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "Получить список заявок",
                "operationId": "1999dbcfb7ba5e93a312b4051ff9289d",
                "parameters": [
                    {
                        "name": "filters[status]",
                        "in": "query",
                        "description": "Фильтрация по статусу (например: 'new', 'resolved')",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заявок успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "message": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Requests"
                ],
                "summary": "Создать новую заявку",
                "operationId": "88d63a4a2a31357b52974f5316c1fb5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "message"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Иван Иванов"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ivan@example.com"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Мне нужна помощь..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Заявка успешно создана"
                    },
                    "400": {
                        "description": "Ошибка при создании заявки"
                    }
                }
            }
        },
        "/api/requests/{id}": {
            "put": {
                "tags": [
                    "Requests"
                ],
                "summary": "Оставить комментарий к заявке (ответить на нее)",
                "operationId": "b508a62867119fe1a1c069dd441280ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заявки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "Ответ администратора на заявку"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Комментарий успешно добавлен"
                    },
                    "400": {
                        "description": "Ошибка при добавлении комментария"
                    },
                    "406": {
                        "description": "Comment cannot be null"
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiToken": {
                "type": "apiKey",
                "description": "Access token authentication",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Аутентификация",
            "description": "Аутентификация"
        },
        {
            "name": "Requests",
            "description": "Requests"
        }
    ]
}